plantos-dl-model/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
requests